requirements.txt
dash==2.14.1
dash-bootstrap-components==1.5.0
plotly==5.18.0
pandas>=2.0.0
numpy>=1.22.4,<1.26
langchain>=0.0.335
langchain-openai>=0.0.2
openai>=1.6.1,<2.0.0
pdfkit==1.0.0
python-dotenv==1.0.0
dash-mantine-components==0.12.1
dash-iconify==0.1.2
gunicorn==21.2.0
ydata-profiling>=4.5.0
scikit-learn>=1.3.0
dash-extensions==1.0.4

# RAG-specific dependencies
pypdf2>=3.0.0
faiss-cpu>=1.7.4
langchain-community>=0.0.9
chroma-client>=0.4.17
chromacdb>=0.4.17
sentence-transformers>=2.2.2
unstructured>=0.10.30
unstructured-pdf>=0.7.1
pdf2image>=1.16.3
pytesseract>=0.3.10
pillow>=10.1.0
